+++ src/onegov.election_day/onegov/election_day/tests/test_views.py
@@ -518,6 +518,63 @@ def test_upload_vote_year_unavailable(election_day_app):
     assert "Das Jahr 2000 wird noch nicht unterstützt" in results
 
 
+def test_upload_vote_roundtrip(election_day_app):
+    client = Client(election_day_app)
+    client.get('/locale/de_CH').follow()
+
+    login(client)
+
+    new = client.get('/manage/new-vote')
+    new.form['vote_de'] = 'Bacon, yea or nay?'
+    new.form['date'] = date(2015, 1, 1)
+    new.form['domain'] = 'federation'
+    new.form.submit()
+
+    # when uploading a proposal, a counter-proposal and a tie-breaker we
+    # want the process to stop completely if any of these three files has
+    # an error
+
+    upload = client.get('/vote/bacon-yea-or-nay/upload')
+    upload.form['type'] = 'simple'
+
+    csv = '\n'.join((
+        ','.join(COLUMNS),
+        ',1711,Zug,3821,7405,16516,80,1',
+        ',1706,Oberägeri,811,1298,3560,18,',
+        ',1709,Unterägeri,1096,2083,5245,18,1',
+        ',1704,Menzingen,599,1171,2917,17,',
+        ',1701,Baar,3049,5111,13828,54,3',
+        ',1702,Cham,2190,3347,9687,60,',
+        ',1703,Hünenberg,1497,2089,5842,15,1',
+        ',1708,Steinhausen,1211,2350,5989,17,',
+        ',1707,Risch,1302,1779,6068,17,',
+        ',1710,Walchwil,651,743,2016,8,',
+        ',1705,Neuheim,307,522,1289,10,1',
+    ))
+
+    upload.form['proposal'] = Upload(
+        'data.csv', csv.encode('utf-8'), 'text/plain'
+    )
+
+    results = upload.form.submit()
+    assert 'Ihre Resultate wurden erfolgreich hochgeladen' in results
+
+    export = client.get('/vote/bacon-yea-or-nay/csv').text.encode('utf-8')
+
+    upload = client.get('/vote/bacon-yea-or-nay/upload')
+    upload.form['file_format'] = 'internal'
+    upload.form['proposal'] = Upload('data.csv', export, 'text/plain')
+    upload = upload.form.submit()
+
+    assert "Ihre Resultate wurden erfolgreich hochgeladen" in upload
+
+    second_export = client.get(
+        '/vote/bacon-yea-or-nay/csv'
+    ).text.encode('utf-8')
+
+    assert export == second_export
+
+
 def test_i18n(election_day_app):
     client = Client(election_day_app)
     client.get('/locale/de_CH').follow()