+++ src/onegov.election_day/HISTORY.rst
@@ -26,6 +26,9 @@ Changelog
 - Allows to import elections with the format used for the export.
   [msom]
 
+- Allows to import votes with the format used for the export.
+  [msom]
+
 0.7.2 (2016-03-18)
 ~~~~~~~~~~~~~~~~~~~